Return the years when survey data available for the site specified by lan, lokalId and fiske

Return the years when survey data available for the site specified by lan, lokalId and fiske

Usage

kul_fangst_ar(lan, lokalId, fiske)

Arguments

lan

character, name of the county (län) where site (lokalId) is located

lokalId

character, id of the site (lokalId) to get years for

fiske

character, type of survey (fiske) to get years for

Value

A vector of integers with the years when survey data available for the site (lokalId)

See also

Examples

kul_fangst_ar("Blekinge län", 204, "Yngelinventering")
#> [1] 2015 2014 2013 2012 2011 2010 2009 2008