Return counties (län) known in the KUL database

Return counties (län) known in the KUL database

Usage

kul_fangst_lan()

Value

A vector of type character containing names of counties (län) known in the KUL database

See also

Examples

# This example will access the Internet to get the data and will not work offline
counties <- kul_fangst_lan()
counties
#>  [1] "Blekinge län"                    "Blekinge län/Skåne län"         
#>  [3] "Ej svenskt vatten"               "Gotlands län"                   
#>  [5] "Gävleborgs län"                  "Gävleborgs län, Uppsala län"    
#>  [7] "Hallands län"                    "Kalmar län"                     
#>  [9] "Kalmar/Blekinge/Skåne län"       "Landskapet Åland"               
#> [11] "Norrbotten"                      "Norrbottens län"                
#> [13] "Norrbottens/Västerbottens län"   "Skåne län"                      
#> [15] "Stockholms län"                  "Stockholms/Södermanlands län"   
#> [17] "Södermanlands län"               "Södermanlands/Östergötlands län"
#> [19] "Uppsala län"                     "Uppsala/Stockholms län"         
#> [21] "Utsjögrund"                      "Västerbottens län"              
#> [23] "Västernorrland"                  "Västernorrlands län"            
#> [25] "Västernorrlands/Gävleborgs län"  "Västra Götalands län"           
#> [27] "Östergötlands län"