Return the types of surveys available for the site specified by lan and lokalId

Return the types of surveys available for the site specified by lan and lokalId

Usage

kul_individ_fisken(lan, lokalId)

Arguments

lan

character, name of the county (län) where site (lokalId) is located

lokalId

integer, id of the site (lokalId) to get types of surveys for

Value

A vector of characters with the types of surveys available for the site (lokalId)

See also

Examples

# This example will access the Internet to get the data and will not work offline
kul_individ_fisken("Blekinge län", 204)
#> [1] "Gåsöfjärden inventering"          "Inventering Nordiska nät augusti"
#> [3] "Inventeringsfiske(LNU)"           "Utlängan-Mellanskär"