Get aggregated data from the Nors API

This function returns electrofishing data from the Nors API. You can filter the data by county (Lan), main drainage area (HaroNr) or lake (Sjo). The function does not check if the parameters are valid and it is possible to give combinations that does not make sense, e.g. a haroNr and Lan that does not intersect. It is also possible send requests that return a lot of data, so be careful.

Usage

nors_aggregerad_rapport(Lan = NULL, HaroNr = NULL, Sjo = NULL)

Arguments

Lan

character string with the name of a county (län) in Sweden

HaroNr

integer with the main drainage area (huvudavrinningsområde) number

Sjo

character string with the name a lake (sjö) in Sweden

Value

A data frame with the electrofishing data. There are 44 columns in the data frame consult Nors-documentation.

See also

Examples

# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 24000 (Rickleån).
data <- nors_aggregerad_rapport(HaroNr = 24000)
head(data)
#>            län haroNr     haro nationelltÖvervakningsstationsID           eU_CD
#> 1 Västerbotten  24000 Rickleån                           188337 SE715804-172509
#> 2 Västerbotten  24000 Rickleån                           188344 SE716011-170342
#> 3 Västerbotten  24000 Rickleån                           204014                
#> 4 Västerbotten  24000 Rickleån                           209961 SE714078-173836
#> 5 Västerbotten  24000 Rickleån                           210063 SE716559-169976
#> 6 Västerbotten  24000 Rickleån                           257227 SE718150-168580
#>                           sjö     lokal sweref99N sweref99E area maxDjup höH
#> 1       715804-172509 Göksjön Hela sjön   7158909    763635 1066      20 134
#> 2    716011-170342 Vittjärnen Hela sjön   7160703    741945    5       7 256
#> 3 716840-170625 Sultenträsket Hela sjön   7169034    744658   25       5 238
#> 4      714078-173836 Tryssjön Hela sjön   7141822    777121  120      15 140
#> 5  716559-169976 Lubboträsket Hela sjön   7166135    738216  338      10 250
#> 6    718150-168580 Vitträsket Hela sjön   7181841    724004  198      21 275
#>   antalFisken                                                  fångadeArter
#> 1           1 Abborre,Benlöja,Braxen,Gers,Gädda,Lake,Mört,Nors,Siklöja,Stäm
#> 2           2                                                 Abborre,Öring
#> 3           1                                Abborre,Braxen,Gers,Gädda,Mört
#> 4           2                             Abborre,Gers,Gädda,Harr,Lake,Mört
#> 5           1                                        Abborre,Gädda,Mört,Sik
#> 6           1                                             Abborre,Gädda,Sik
#>   förstaFiskeår senasteFiskeår allaFiskeår  allaEQR8 senasteMetod bNätTyp
#> 1          2014           2014        2014      0.46        Stand Bnord12
#> 2          2010           2024   2010,2024                  Inven Bnord12
#> 3          1993           1993        1993                  Oklas Bdrot14
#> 4          2009           2021   2009,2021 0.46,0.30        Stand Bnord12
#> 5          1993           1993        1993                  Oklas Bdrot14
#> 6          1999           1999        1999                  Inven Bnord12
#>   npueTot wpueTot vanlArt vanlArtWProc nästVanlArt nästVanlArtWProc     eqR8
#> 1   27.93 1428.18 Abborre         71.2        Mört             10.0 0,46 (M)
#> 2   59.00 2433.50 Abborre         76.8       Öring             23.2         
#> 3   13.75 2057.88 Abborre         42.5      Braxen             31.6         
#> 4   26.33 2539.40 Abborre         62.8       Gädda             31.9 0,30 (M)
#> 5   31.25 2309.38 Abborre         55.8        Mört             21.3         
#> 6    7.75  263.44 Abborre         71.3         Sik             28.6         
#>   aindexW5 eindexW3 nMört_100 nMört101_110 nMört111_120 nMört121_130
#> 1 0,45 (H) 0,58 (H)         5            6           25           26
#> 2                           0            0            0            0
#> 3                           1            0            0            2
#> 4 0,29 (G) 0,68 (H)        49           14           12            4
#> 5                           9            6            3            3
#> 6                           0            0            0            0
#>   nMört131_140 nMört141_150 npue_3 npue3_6 npue6_12 npue12_20 npue20_35
#> 1           27           17  48.20   43.82    12.56         4       0.0
#> 2            0            0  92.50   50.00     1.00         0       0.0
#> 3            0            0  25.00    7.00     0.00         0       0.0
#> 4            7            8  47.80   33.83    11.15         0       0.0
#> 5            7            7  29.50   38.67    16.00         0       0.0
#> 6            0            0  14.67    8.63    10.00         0       0.5
#>   npue35_50 npue50_75 npue75_
#> 1         0         0       0
#> 2         0         0       0
#> 3         0         0       0
#> 4         0         0       0
#> 5         0         0       0
#> 6         0         0       0
#>                                                                url
#> 1 http://aquarapport.slu.se/default.aspx?ID=99&X=7158040&Y=1725090
#> 2 http://aquarapport.slu.se/default.aspx?ID=99&X=7160110&Y=1703420
#> 3 http://aquarapport.slu.se/default.aspx?ID=99&X=7168400&Y=1706250
#> 4 http://aquarapport.slu.se/default.aspx?ID=99&X=7140780&Y=1738360
#> 5 http://aquarapport.slu.se/default.aspx?ID=99&X=7165590&Y=1699760
#> 6 http://aquarapport.slu.se/default.aspx?ID=99&X=7181500&Y=1685800
dim(data)
#> [1] 12 44