Get individ data from the Nors API

This function returns electrofishing data from the Nors API. You can filter the data by county (Lan), main drainage area (HaroNr) or lake (Sjo). The function does not check if the parameters are valid and it is possible to give combinations that does not make sense, e.g. a haroNr and Lan that does not intersect. It is also possible send requests that return a lot of data, so be careful.

Usage

nors_individ_rapport(lan, kommun = NULL, vatten = NULL)

Arguments

lan

character string with the name of a county (län) in Sweden

kommun

character string with the name of a municipality in county (lan) in Sweden

vatten

character string with the name a lake (vatten) in Sweden

Value

A data frame with survey data. There are 7 columns in the data frame consult NORS-documentation.

See also

Examples

# This example will access the Internet to get the data and will not work offline
# Get all data from lake "626842-147473 Flaken" in county "Kalmar Län".
data <- nors_individ_rapport("Kalmar Län", vatten = "626842-147473 Flaken")
head(data)
#>          län               vatten      datum     lokal nätnummer     art längd
#> 1 Kalmar Län 626842-147473 Flaken 1984-06-12 Hela sjön         1 Abborre    75
#> 2 Kalmar Län 626842-147473 Flaken 1984-06-12 Hela sjön         1 Abborre    75
#> 3 Kalmar Län 626842-147473 Flaken 1984-06-12 Hela sjön         1 Abborre    75
#> 4 Kalmar Län 626842-147473 Flaken 1984-06-12 Hela sjön         1 Abborre    85
#> 5 Kalmar Län 626842-147473 Flaken 1984-06-12 Hela sjön         1 Abborre    85
#> 6 Kalmar Län 626842-147473 Flaken 1984-06-12 Hela sjön         1 Abborre    85
dim(data)
#> [1] 5396    7