Returns lakes (vatten) in a county (and optionally municipality )in Sweden.

Returns lakes (vatten) in a county (and optionally municipality )in Sweden.

Usage

nors_individ_vatten(lan, kommun = NULL)

Arguments

lan

character string with the name of a county (län) in Sweden

kommun

character string with the name of a municipality in county (lan) in Sweden

Value

A list of lakes (vatten) in specified county (län).

See also

Examples

# This example will access the Internet to get the data and will not work offline
nors_individ_vatten(lan = "Uppsala Län", kommun = "Älvkarleby")
#> [1] "672632-158939 Dalälven" "671895-158664 Djupsjön"