Get electrofishing data from the SERS API

This function returns electrofishing data from the SERS API. You can filter the data by county (lan), municipality (kommun), main drainage area (haroNr). The function does not check if the parameters are valid and it is possible to give combinations that does not make sense, e.g. a haroNr and Lan that does not intersect. It is also possible send requests that return a lot of data, so be careful.

Usage

sers_aggregerad_rapport(Lan = NULL, haroNr = NULL, Vattendrag = NULL)

Arguments

Lan

character string with the name of a county (län) in Sweden

haroNr

integer with the main drainage area (huvudavrinningsområde) number

Vattendrag

character string with the name a river (vattendrag) in Sweden

Value

A data frame with the electrofishing data. There are 30 columns in the data frame consult SERS-documentation.

See also

Examples

# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 24000 (Rickleån).
data <- sers_aggregerad_rapport(haroNr = 24000)
head(data)
#>            län hflodomr huvudflodområde      vattendrag            lokal
#> 1 Västerbotten    24000        Rickleån       Aggbäcken   Aggbäcken övre
#> 2 Västerbotten    24000        Rickleån       Aggbäcken       Aggån bron
#> 3 Västerbotten    24000        Rickleån      Frödbäcken Nedströms ödehus
#> 4 Västerbotten    24000        Rickleån Holktjärnbäcken Jan-Pettersmyran
#> 5 Västerbotten    24000        Rickleån Holktjärnbäcken        Mynningen
#> 6 Västerbotten    24000        Rickleån      Holmbäcken      Ovan stugan
#>         lokalID nationelltProvplatsID nationelltÖvervakningsstationsID
#> 1 715736-170890                367443                               -9
#> 2 715880-170913                367447                               -9
#> 3 717326-171264                    NA                               -9
#> 4 715515-171440                115705                           209939
#> 5 715600-171470                115708                           233548
#> 6 716084-171546                115727                           209944
#>             eU_CD      mS_CD svaR_Ver sweref99N sweref99E antalFisken
#> 1 SE715594-747363 WA65121662 SVAR2016   7158028    747463           1
#> 2 SE715848-170926 WA65121662 SVAR2016   7159474    747675           1
#> 3 NW717460-750582 WA21808797 SVAR2016   7173968    750995           1
#> 4 SE715498-171424 WA65492019 SVAR2016   7155883    752984          34
#> 5                 WA65492019            7156737    753274          31
#> 6 SE716064-171647 WA64148253 SVAR2016   7161585    753972          21
#>   förstaFiskeår senasteFiskeår
#> 1          2021           2021
#> 2          2021           2021
#> 3          2024           2024
#> 4          1990           2024
#> 5          1991           2024
#> 6          2003           2024
#>                                                                                                                                                                     fiskeår
#> 1                                                                                                                                                                      2021
#> 2                                                                                                                                                                      2021
#> 3                                                                                                                                                                      2024
#> 4 1990,1991,1992,1993,1994,1995,1996,1997,1998,1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019,2020,2021,2022,2023,2024
#> 5                1991,1992,1993,1994,1995,1996,1997,1998,1999,2000,2001,2002,2004,2005,2006,2007,2008,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019,2020,2022,2023,2024
#> 6                                                                  2003,2004,2005,2006,2007,2008,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019,2020,2021,2022,2023,2024
#>   avfiskadYta               arter öringLägstaTäthet öringHögstaTäthet
#> 1        63.0               Gädda               0.0               0.0
#> 2        80.0       Abborre,Gädda               0.0               0.0
#> 3        18.0        Ingen fångst               0.0               0.0
#> 4       138.1 Elritsa,Gädda,Öring               0.0             137.8
#> 5       142.8       Elritsa,Öring               0.0              45.4
#> 6        64.3       Elritsa,Öring               2.5              51.7
#>   öring0LägstaTäthet öring0HögstaTäthet predikteradTäthetLaxfisk
#> 1                  0                0.0                      9.1
#> 2                  0                0.0                      4.9
#> 3                  0                0.0                     15.1
#> 4                  0              110.3                     16.9
#> 5                  0               40.4                     10.7
#> 6                  0               46.9                     13.6
#>   andelMedÖring0Proc andelMedÖringStörreProc vattenhastighet senasteVIX
#> 1               0.00                    0.00           Ström   0,24 (O)
#> 2               0.00                    0.00           Ström   0,01 (D)
#> 3               0.00                    0.00            Lugn   0,00 (D)
#> 4               0.74                    0.82           Ström   0,44 (M)
#> 5               0.55                    0.90           Ström   0,47 (G)
#> 6               0.71                    1.00           Ström   0,30 (M)
#>                                                                                                                                                                     allaVIX
#> 1                                                                                                                                                                      0.24
#> 2                                                                                                                                                                      0.01
#> 3                                                                                                                                                                      0.00
#> 4 0.00,0.00,0.00,0.12,0.12,0.44,0.23,0.22,0.00,0.36,0.41,0.44,0.40,0.46,0.32,0.46,0.39,0.44,0.44,0.35,0.41,0.40,0.44,0.43,0.40,0.43,0.41,0.44,0.39,0.41,0.38,0.41,0.46,0.44
#> 5                0.24,0.36,0.00,0.00,0.22,0.00,0.41,0.33,0.32,0.40,0.29,0.49,0.33,0.44,0.32,0.47,0.45,0.32,0.33,0.36,0.38,0.45,0.29,0.40,0.27,0.36,0.39,0.23,0.40,0.45,0.47
#> 6                                                                  0.22,0.43,0.21,0.40,0.41,0.30,0.35,0.45,0.45,0.47,0.48,0.49,0.41,0.44,0.41,0.34,0.25,0.39,0.50,0.48,0.30
#>                                                                url
#> 1 http://aquarapport.slu.se/default.aspx?ID=98&X=7157365&Y=1708905
#> 2 http://aquarapport.slu.se/default.aspx?ID=98&X=7158809&Y=1709136
#> 3 http://aquarapport.slu.se/default.aspx?ID=98&X=7173262&Y=1712640
#> 4 http://aquarapport.slu.se/default.aspx?ID=98&X=7155150&Y=1714400
#> 5 http://aquarapport.slu.se/default.aspx?ID=98&X=7156000&Y=1714700
#> 6 http://aquarapport.slu.se/default.aspx?ID=98&X=7160840&Y=1715460
dim(data)
#> [1] 91 30