Get the municipalities (kommuner) in a county (län)

This function returns the municipalities (kommuner) in a county (län) in Sweden. The parameter lan must be a valid county (län) in Sweden.

sers_vix_kommuner(lan)

Arguments

lan

A character string with the name of a county (län) in Sweden

Value

A vector of type character containing names of municipalities (kommuner) in the county (län)

See also

Examples

# This example will access the Internet to get the data and will not work offline
kommuner <- sers_vix_kommuner("Västra Götaland")
kommuner
#>  [1] "Ale"         "Alingsås"    "Bengtsfors"  "Bollebygd"   "Borås"      
#>  [6] "Dals-Ed"     "Essunga"     "Falköping"   "Färgelanda"  "Grästorp"   
#> [11] "Gullspång"   "Göteborg"    "Götene"      "Herrljunga"  "Hjo"        
#> [16] "Härryda"     "Karlsborg"   "Kungälv"     "Lerum"       "Lidköping"  
#> [21] "Lilla Edet"  "Lysekil"     "Mariestad"   "Mark"        "Mellerud"   
#> [26] "Mullsjö"     "Munkedal"    "Mölndal"     "Orust"       "Partille"   
#> [31] "Skara"       "Skövde"      "Sotenäs"     "Stenungsund" "Strömstad"  
#> [36] "Svenljunga"  "Tanum"       "Tibro"       "Tidaholm"    "Tjörn"      
#> [41] "Tranemo"     "Trollhättan" "Töreboda"    "Uddevalla"   "Ulricehamn" 
#> [46] "Vara"        "Vårgårda"    "Vänersborg"  "Åmål"