dvfisk::sites A data frame with information about electrofishing sites in SERS

Data from the Swedish Electrofishing Register (SERS) containing information about electrofishing sites. The documentation is based on the Excel file provided by SLU and avaliable in Swedish only.

Data frame dvfisk::sites is a list of sites defined in SERS with 23 variables described below.

dvfisk::sites_sf is a simple feature points object for the sites in dvfisk::sites use dplyr::left_join(sites, sites_sf, by = dplyr::join_by(XKOORLOK, YKOORLOK)) to join the data frames. You might want to use library(sf) in your script to work with the simple feature object.

sites

sites_sf

Format

Columns in dvfisk::sites are:

lan

Länen kodas 1 till 25 (Stockholms till Norrbottens län).

hflodomr

Huvudflodområdesnummer enligt SMHI:s numrering (Vattendragsregistret).

biflnr

Biflödesnummer enligt SMHI (Vattendragsregistret).

vdragnam

Vattendragets namn.

xkoorvdr

Vattendragets mynningskoordinat i X-led (nord-syd) enligt SMHI (Vattendragsregistret).

ykoorvdr

Vattendragets mynningskoordinat i Y-led (väst-öst) enligt SMHI (Vattendragsregistret).

lokalnam

Elfiskelokalens namn

lokalnr

Elfiskelokalens nummer (om blankt saknas lokalnummer för aktuell elfiskelokal).

xkoorlok

X-koordinat (nord-syd; sexsiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.

ykoorlok

Y-koordinat (väst-öst; sexsiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.

hoh

Höjd över havet i meter.

avriomrk

Avrinningsområdets storlek uppströms elfiskelokalen (km2). Anges enbart i fyra klasser (<10, <100, <1000 resp >1000).

andsjopr

Andelen sjöar i % av avrinningsområdet uppströms elfiskelokalen. Anges enbart i fyra klasser (<1, <5, <10, >10).

avstupp

Avstånd (km) till uppströms sjö eller större sel (lugnparti) om minst 1 hektar i huvudfåran.

avstner

Avstånd (km) till nedströms sjö eller större sel (lugnparti) om minst 1 hektar i huvudfåran.

kommunnr

Anger vilken kommun som elfiskelokalen ligger i enligt SCB.

kommun

Namn på kommunen där elfiskelokalen är lokaliserad.

topokart

Här anges vilken topografisk karta elfiskelokalen ligger på, ex 10B SO. Lämnas tomt om ej angivet.

x_rt90_sju

X-koordinat (nord-syd; sjusiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.

y_rt90_sju

Y-koordinat (väst-öst; sjusiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.

s99tm_n

Koordinat (nord-syd, Sweref99 TM) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!

s99tm_e

Koordinat (väst-öst, Sweref99 TM) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!

ddlat

Latitud för elfiskelokalen (WGS84) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!

ddlong

Latitud för elfiskelokalen (WGS84) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!

An object of class sf (inherits from tbl_df, tbl, data.frame) with 21749 rows and 3 columns.

Source

Swedish Electrofishing Register (SERS) https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-testfishing-in-streams/