Get electrofishing data from the SERS API

This function returns electrofishing data from the SERS API. You can filter the data by county (lan), municipality (kommun), main drainage area (haroNr), start date (startdatum) and end date (slutdatum). You must provide at least one of the parameters. The function does not check if the parameters are valid and it is possible to give combinations that does not make sense, e.g. a haroNr and lan that does not intersect. It is also possible send requests that return a lot of data, so be careful.

Usage

sers_vix_rapport(
  lan = NULL,
  kommun = NULL,
  haroNr = NULL,
  startdatum = NULL,
  slutdatum = NULL
)

Arguments

lan

character string with the name of a county (län) in Sweden

kommun

character string with the name of a municipality (kommun) in Sweden

haroNr

integer with the main drainage area (huvudavrinningsområde) number

startdatum

character string with the start date in the format "YYYY-MM-DD"

slutdatum

character string with the end date in the format "YYYY-MM-DD"

Value

A data frame with the electrofishing data. There are 244 columns in the data frame consult SERS-documentation.

See also

Examples

# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 86000 (Mörrumsån) in 2021
data <- sers_vix_rapport(haroNr = 86000, startdatum = "2021-01-01", slutdatum = "2021-12-31")
head(data)
#>        län länskod harO_NR huvudavrinningsområde hflodomr biflnr    vattendrag
#> 1 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 2 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 3 Blekinge      10   86000             Mörrumsån      086      1 Ällhölabäcken
#> 4 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 5 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 6 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#>   xkoorvdr ykoorvdr                lokal lokalnr  rT90_X  rT90_Y swereF99_TM_N
#> 1   622563   143423 Åkeholm stenbro nedr         6240950 1434300       6238608
#> 2   622563   143423  Knaggelid utanfsido       1 6241520 1434170       6239176
#> 3   623661   143643       Ned gamla bron         6239469 1436907       6237158
#> 4   622563   143423  Åkeholm-ön sidofåra         6240850 1434500       6238511
#> 5   622563   143423            Knaggelid         6242100 1434050       6239755
#> 6   622563   143423      Knaggelid vägen       2 6242340 1433890       6239993
#>   swereF99_TM_E fiskedatum  syfte artantal abbor asp benlö bjöna besim stesi
#> 1        484167   20210910 Nö-lax        8     0   0   4.4     0   0.6     0
#> 2        484030   20210927 Nö-lax        6     0   0   0.0     0   3.5     0
#> 3        486790   20210930   Rkeu        0     0   0   0.0     0   0.0     0
#> 4        484368   20210902 Nö-lax        6     0   0   3.6     0   2.5     0
#> 5        483904   20210929 Nö-lax        4     0   0   0.0     0   3.2     0
#> 6        483741   20210831 Nö-lax        1     0   0   0.0     0   0.0     0
#>   simpa hosim braxe bäcne flone havne nejonöga bäcrö0 bäcrö bäcrötot bäcrömax
#> 1     0     0     0     0     0     0        0      0     0        0       NA
#> 2     0     0     0     0     0     0        0      0     0        0       NA
#> 3     0     0     0     0     0     0        0      0     0        0       NA
#> 4     0     0     0     0     0     0        0      0     0        0       NA
#> 5     0     0     0     0     0     0        0      0     0        0       NA
#> 6     0     0     0     0     0     0        0      0     0        0       NA
#>   bäcrömin bäCRmax0 bäcröxö0 bäcröxör bäcröxöt bäCXömax bäCXömin bäCXömx0 elrit
#> 1       NA       NA        0        0        0       NA       NA       NA  67.5
#> 2       NA       NA        0        0        0       NA       NA       NA  83.5
#> 3       NA       NA        0        0        0       NA       NA       NA   0.0
#> 4       NA       NA        0        0        0       NA       NA       NA  14.6
#> 5       NA       NA        0        0        0       NA       NA       NA  19.6
#> 6       NA       NA        0        0        0       NA       NA       NA   0.0
#>   flodkräf signkräf kräfta färna gers grolö grönl gädda gös harr0 harr harrtot
#> 1        0      1.5      0   6.3    0     0     0     0   0     0    0       0
#> 2        0      0.0      0   2.3    0     0     0     0   0     0    0       0
#> 3        0      0.0      0   0.0    0     0     0     0   0     0    0       0
#> 4        0      0.0      0   1.3    0     0     0     0   0     0    0       0
#> 5        0      0.0      0   0.0    0     0     0     0   0     0    0       0
#> 6        0      0.0      0   0.0    0     0     0     0   0     0    0       0
#>   harrmax harrmin harrmax0 id kanrö0 kanrö kanrötot lake  lax0  lax laxtot
#> 1      NA      NA       NA  0      0     0        0    0 186.3 14.4  200.7
#> 2      NA      NA       NA  0      0     0        0    0  46.0  5.1   51.1
#> 3      NA      NA       NA  0      0     0        0    0   0.0  0.0    0.0
#> 4      NA      NA       NA  0      0     0        0    0 130.3 24.2  154.5
#> 5      NA      NA       NA  0      0     0        0    0  46.8 13.0   59.8
#> 6      NA      NA       NA  0      0     0        0    0   0.0  0.0    0.0
#>   laxmax laxmin laxmax0 lax0ber laxber lax0pval laxpval laxfisk0 laxfisk
#> 1    130     45      85       3      3     0.80    0.96        0       0
#> 2    130     50      90       3      1     0.88    0.91        0       0
#> 3     NA     NA      NA      NA     NA       NA    0.00        0       0
#> 4    130     10      75       3      3     0.87    0.86        0       0
#> 5    135     55      85       1      1     0.45    0.55        0       0
#> 6     NA     NA      NA      NA     NA       NA    0.00        0       0
#>   laxfixto laxfimax laxfimin laxfmax0 laxör0 laxör laxörTOT laxörmax laxörmin
#> 1        0       NA       NA       NA      0     0        0       NA       NA
#> 2        0       NA       NA       NA      0     0        0       NA       NA
#> 3        0       NA       NA       NA      0     0        0       NA       NA
#> 4        0       NA       NA       NA      0     0        0       NA       NA
#> 5        0       NA       NA       NA      0     0        0       NA       NA
#> 6        0       NA       NA       NA      0     0        0       NA       NA
#>   mal karp mört nissö nors regnb0 regnb regnbtot regnmax0 rödin0 rödin rödintot
#> 1   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 2   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 3   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 4   0    0  0.4     0    0      0     0        0       NA      0     0        0
#> 5   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 6   0    0  0.0     0    0      0     0        0       NA      0     0        0
#>   rödinmax rödinmin rödimax0 ruda sankr sarv sik siklö skrub rödsp småsp spigg
#> 1       NA       NA       NA    0   0.0    0   0     0     0     0     0     0
#> 2       NA       NA       NA    0   1.1    0   0     0     0     0     0     0
#> 3       NA       NA       NA    0   0.0    0   0     0     0     0     0     0
#> 4       NA       NA       NA    0   0.0    0   0     0     0     0     0     0
#> 5       NA       NA       NA    0   0.0    0   0     0     0     0     0     0
#> 6       NA       NA       NA    0   0.0    0   0     0     0     0     0     0
#>   stosp stäm sutar vimma  ål ålmax ålmin öring0 öring öringtot öringmax
#> 1     0    0     0     0 1.6   200    70    3.6   0.7      4.3      165
#> 2     0    0     0     0 0.0    NA    NA    6.7   0.0      6.7       95
#> 3     0    0     0     0 0.0    NA    NA    0.0   0.0      0.0       NA
#> 4     0    0     0     0 0.0    NA    NA    0.0   0.0      0.0       NA
#> 5     0    0     0     0 0.0    NA    NA    3.0   0.0      3.0       90
#> 6     0    0     0     0 0.0    NA    NA    0.0   0.0      0.0       NA
#>   öringmin örinmax0 örin0ber öringber öri0pval örinpval sanst lerst smsmörb
#> 1       70       80        1        1     0.86     0.91     0     0       0
#> 2       75       95        3       -9     0.68     0.00     0     0       0
#> 3       NA       NA       NA       NA     0.00     0.00     0     0       0
#> 4       NA       NA       NA       NA     0.00     0.00     0     0       0
#> 5       65       90        1       -9     0.48     0.00     0     0       0
#> 6       NA       NA       NA       NA     0.00     0.00     0     0       0
#>   utförare ansvarig fabrikat aggregat volt strstyrk pulsfrek antutfis metod
#> 1   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 2   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 3    Lst10    Lst10  Biowave  Batteri  580      0.2      400        1  Kval
#> 4   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 5   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        1  Kval
#> 6   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        1  Kval
#>   bredd langd lokalbre area avfiskbr medbredd medyta grumligh vtnfarg maxdjup
#> 1  25.0  32.0      5.0  160        N       -9     -9        1       2    0.65
#> 2  25.0  15.0      5.8   87        N       -9     -9        1       2    0.60
#> 3   1.4  20.0      1.4   27                -9     -9        1       2    0.30
#> 4  25.0  23.5     10.0  235        N       -9     -9        1       2    0.60
#> 5  35.0  22.0      9.5  209        N       -9     -9        1       2    0.90
#> 6  25.0  27.0      2.0   54        N       -9     -9        1       2    1.10
#>   medeldju bottento substr1 substr2 substr3 vattenni vattenha vattente lufttemp
#> 1     0.55        2   Sten2   Sten1  Block1      Låg     Strö     17.6      8.0
#> 2     0.35        2  Block1   Sten2   Sten1      Låg     Strö     13.8     14.1
#> 3     0.20        2   Sten1    Grus   Sten2      Låg     Strö     13.0     12.0
#> 4     0.35        2   Sten2   Sten1  Block2      Låg     Strö     17.2     16.0
#> 5     0.60       -9  Block1   Sten2  Block2      Låg     Strå     13.6     13.0
#> 6     0.90        1   Sten2  Block1   Sten1      Låg     Lugn     16.6     17.0
#>   ovegmang uvegmang uvegtyp narmiljo domtra1 domtra2 beskuggn vedivant vedivpar
#> 1              Mått    Palg    Artif      Al               10       -9     -9.0
#> 2              Mått    Moss    Lövsk      Al     Ask        0       -9     -9.0
#> 3     Sakn     Ring    Moss    Lövsk      Al    Lönn       90        0      0.0
#> 4              Mått    Moss      Äng      Al               30        0      0.0
#> 5              Mått    Moss    Blask      Al                0        3      1.4
#> 6     Ring     Mått    Blom    Lövsk      Al   Björk       20        0      0.0
#>   vandhind typavpop lokalvar kalkpave kalkdatu kalkntyp paverkt1 paverkt2
#> 1                         -9                 0                           
#> 2      Upp    Vandr        1       Ja       -9             Flott         
#> 3      Ned    Ström       -9                 0             Torka         
#> 4     Inga    Vandr        1       Ja       -9             Flott         
#> 5      Upp    Vandr        2       Ja       -9                           
#> 6      Upp    Vandr        0       Ja       -9      Bla                  
#>   paverkt3 pastyrk1 pastyrk2 pastyrk3                  elfiske_anmarkni
#> 1                -9       -9       -9                                  
#> 2                 1       -9       -9                                  
#> 3                 2       -9       -9 5 % av ytan avräknad av utföraren
#> 4                 1       -9       -9                                  
#> 5                -9       -9       -9                                  
#> 6                -9       -9       -9                                  
#>   xkoorlok ykoorlok wgS84_Dec_N wgS84_Dec_E provplatsID stationsID
#> 1   624095   143430    56.29200    14.74421      107160     203968
#> 2   624152   143417    56.29710    14.74197      167489     234212
#> 3   623946   143690    56.27906    14.78666      363965     363964
#> 4   624085   143450    56.29113    14.74747      181205     247382
#> 5   624210   143405    56.30229    14.73989      107181     204506
#> 6   624234   143389    56.30443    14.73724      167502     234215
#>             eU_CD      mS_CD svaR_Ver kommunnr    kommun   år månad vtyP_ED
#> 1                 WA96144336              1082 Karlshamn 2021     9     Hav
#> 2                 WA96144336              1082 Karlshamn 2021     9     Hav
#> 3 SE623453-486436 WA26415463 SVAR2016     1082 Karlshamn 2021     9        
#> 4 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     9     Hav
#> 5 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     9     Hav
#> 6 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     8     Hav
#>   vtyp_123 hoh avriomrk andsjopr avstupp avstner mindistsj lutninG_PROM
#> 1        3  50   <10000      >10    10.0     1.3       1.3          5.0
#> 2        3  52   <10000      >10    10.0     1.8       1.8          5.0
#> 3        1  48     <100      <10     1.5     1.7       1.5          5.2
#> 4        3  49   <10000      >10    10.0     1.0       1.0          1.0
#> 5        3  56   <10000      >10    10.0     2.5       2.5          3.6
#> 6        3  57   <10000      >10    10.0     2.6       2.6          3.6
#>   medtempar medT_JULI nsp nölax pred_nölax p_VIX_nölax nandtol pred_nandtol
#> 1         7      15.5   7 205.0    31.6649      0.9749  0.0210       0.0289
#> 2         7      15.5   6  57.8    31.6649      0.7341  0.0000       0.0245
#> 3         7      15.5   0   0.0     8.2685      0.0137  0.0000       0.0157
#> 4         7      15.5   6 154.5    42.3714      0.9128  0.0226       0.0330
#> 5         7      15.5   4  62.8    40.4553      0.6767  0.0000       0.0188
#> 6         7      15.5  NA    NA    33.9831          NA      NA       0.0211
#>   p_VIX_nandtol nandlith pred_nandlith p_VIX_nandlith spproptol pred_spproptol
#> 1        0.5659   0.9790        0.9601         0.5800    0.2857         0.0745
#> 2        0.8499   0.9926        0.9699         0.6253    0.0000         0.0481
#> 3        0.7964   0.0000        0.8724         0.0000    0.0000         0.0153
#> 4        0.5824   0.9774        0.9509         0.6037    0.3333         0.0972
#> 5        0.8178   1.0000        0.9776         0.7073    0.0000         0.0684
#> 6            NA       NA        0.9783             NA        NA         0.0317
#>   p_VIX_spproptol sppropint pred_sppropintol p_VIX_sppropintol kvot pred_kvot
#> 1          0.0992    0.4286           0.5228              0.40    1    0.2373
#> 2          0.8387    0.5000           0.6052              0.39    1    0.2373
#> 3          0.7107    0.0000           0.8509              0.00   NA    0.6947
#> 4          0.0909    0.3333           0.4228              0.40    1    0.2373
#> 5          0.8818    0.7500           0.5065              0.74    1    0.2373
#> 6              NA        NA           0.6540                NA   NA    0.2373
#>   p_VIX_kvot simpson pred_simpson p_VIXh_simpson  viXsm   viXh    vix viX_klass
#> 1     0.9303  0.4486       0.2998         0.5894 0.7228 0.5822 0.5927         2
#> 2     0.9303  0.5608       0.2470         0.2543 0.6711 0.5402 0.7288         2
#> 3         NA      NA       0.0860             NA 0.0000 0.0000 0.0000         5
#> 4     0.9303  0.2297       0.3695         0.5917 0.7137 0.5671 0.5880         2
#> 5     0.9303  0.4569       0.3637         0.7402 0.7636 0.6178 0.7923         1
#> 6         NA      NA       0.2230             NA 0.0000 0.0000 0.0000         5
#>   fisktom                                  viX_anmärkni viX_anmärkni2
#> 1      NA                                                            
#> 2      NA                                                            
#> 3       1                  Klass 5 pga att fisk saknas.              
#> 4      NA                                                            
#> 5      NA                                                            
#> 6       1 Klass 5 pga att fisk saknas. Kräfta fångades.              
#>   pre_sd_VIX_c    p_1    p_2    p_3    p_4   p_5 p23_diff viX23_gränsfall
#> 1       0.1552 0.1567 0.6347 0.1886 0.0196 5e-04   0.4461              NA
#> 2       0.1625 0.4501 0.4965 0.0509 0.0025 0e+00   0.4456              NA
#> 3       0.1535 0.0000 0.0000 0.0000 0.0000 1e+00   0.0000              NA
#> 4       0.1508 0.1426 0.6467 0.1921 0.0183 4e-04   0.4546              NA
#> 5       0.1436 0.6180 0.3703 0.0116 0.0002 0e+00   0.3587              NA
#> 6       0.1664 0.0000 0.0000 0.0000 0.0000 1e+00   0.0000              NA
#>   gränsfall_string viXmorf påverkat_VIXmorf
#> 1                   0.5917                0
#> 2                   0.5419                0
#> 3                   0.0347                1
#> 4                   0.4924                0
#> 5                   0.6287                0
#> 6                   0.0199                1
#>                                                           komm_VIXmorf
#> 1          God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 2          God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 3 VIX<0,467 och fisktomt elfiskeresultat. VIXmorfberäkning ej pålitlig
#> 4          God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 5          God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 6                   VIX<0,467 och morfologiskt påverkat enligt VIXmorf
dim(data)
#> [1]  60 244