Get electrofishing data from the SERS API

This function returns electrofishing data from the SERS API. You can filter the data by county (lan), municipality (kommun), main drainage area (haroNr), start date (startdatum) and end date (slutdatum). You must provide at least one of the parameters. The function does not check if the parameters are valid and it is possible to give combinations that does not make sense, e.g. a haroNr and lan that does not intersect. It is also possible send requests that return a lot of data, so be careful.

Usage

sers_vix_rapport(
  lan = NULL,
  kommun = NULL,
  haroNr = NULL,
  startdatum = NULL,
  slutdatum = NULL
)

Arguments

lan

character string with the name of a county (län) in Sweden

kommun

character string with the name of a municipality (kommun) in Sweden

haroNr

integer with the main drainage area (huvudavrinningsområde) number

startdatum

character string with the start date in the format "YYYY-MM-DD"

slutdatum

character string with the end date in the format "YYYY-MM-DD"

Value

A data frame with the electrofishing data. There are 244 columns in the data frame consult SERS-documentation.

See also

Examples

# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 86000 (Mörrumsån) in 2021
data <- sers_vix_rapport(haroNr = 86000, startdatum = "2021-01-01", slutdatum = "2021-12-31")
head(data)
#>        län länskod harO_NR huvudavrinningsområde hflodomr biflnr
#> 1 Blekinge      10   86000             Mörrumsån      086      0
#> 2 Blekinge      10   86000             Mörrumsån      086      0
#> 3 Blekinge      10   86000             Mörrumsån      086      0
#> 4 Blekinge      10   86000             Mörrumsån      086      0
#> 5 Blekinge      10   86000             Mörrumsån      086      0
#> 6 Blekinge      10   86000             Mörrumsån      086      0
#>           vattendrag xkoorvdr ykoorvdr               lokal lokalnr  rT90_X
#> 1          Mörrumsån   622563   143423           Knaggelid         6242100
#> 2          Mörrumsån   622563   143423     Knaggelid vägen       2 6242340
#> 3          Mörrumsån   622563   143423 Knaggelid huvudfåra         6243540
#> 4          Mörrumsån   622563   143423 Vittskövle hängbron         6232950
#> 5 Mörrumsån sidofåra   622563   143423 Vittskövle sidofåra         6233040
#> 6 Mörrumsån sidofåra   622563   143423   Vittskövle, kvill         6233120
#>    rT90_Y swereF99_TM_N swereF99_TM_E fiskedatum  syfte artantal abbor asp
#> 1 1434050       6239755        483904   20210929 Nö-lax        4     0   0
#> 2 1433890       6239993        483741   20210831 Nö-lax        1     0   0
#> 3 1433390       6241186        483227   20210913 Nö-lax        6     0   0
#> 4 1435050       6230621        485010   20210903 Nö-lax        4     0   0
#> 5 1435220       6230713        485179   20210801   Uppf        4     0   0
#> 6 1435330       6230794        485288   20210803   Uppf        4     0   0
#>   benlö bjöna besim stesi simpa hosim braxe bäcne flone havne nejonöga bäcrö0
#> 1   0.0     0   3.2     0     0     0     0     0     0     0      0.0      0
#> 2   0.0     0   0.0     0     0     0     0     0     0     0      0.0      0
#> 3   0.0     0  68.8     0     0     0     0     0     0     0      0.0      0
#> 4   0.6     0   0.0     0     0     0     0     0     0     0      0.0      0
#> 5   0.0     0   0.0     0     0     0     0     0     0     0      0.5      0
#> 6   0.0     0   0.0     0     0     0     0     0     0     0      0.0      0
#>   bäcrö bäcrötot bäcrömax bäcrömin bäCRmax0 bäcröxö0 bäcröxör bäcröxöt bäCXömax
#> 1     0        0       NA       NA       NA        0        0        0       NA
#> 2     0        0       NA       NA       NA        0        0        0       NA
#> 3     0        0       NA       NA       NA        0        0        0       NA
#> 4     0        0       NA       NA       NA        0        0        0       NA
#> 5     0        0       NA       NA       NA        0        0        0       NA
#> 6     0        0       NA       NA       NA        0        0        0       NA
#>   bäCXömin bäCXömx0 elrit flodkräf signkräf kräfta färna gers grolö grönl gädda
#> 1       NA       NA  19.6        0        0      0   0.0    0     0     0     0
#> 2       NA       NA   0.0        0        0      0   0.0    0     0     0     0
#> 3       NA       NA 201.6        0        0      0  18.4    0     0     0     0
#> 4       NA       NA   9.1        0        0      0   0.0    0     0     0     0
#> 5       NA       NA  28.6        0        0      0   0.0    0     0     0     0
#> 6       NA       NA  45.8        0        0      0   0.0    0     0     0     0
#>   gös harr0 harr harrtot harrmax harrmin harrmax0 id kanrö0 kanrö kanrötot lake
#> 1   0     0    0       0      NA      NA       NA  0      0     0        0    0
#> 2   0     0    0       0      NA      NA       NA  0      0     0        0    0
#> 3   0     0    0       0      NA      NA       NA  0      0     0        0    0
#> 4   0     0    0       0      NA      NA       NA  0      0     0        0    0
#> 5   0     0    0       0      NA      NA       NA  0      0     0        0    0
#> 6   0     0    0       0      NA      NA       NA  0      0     0        0    0
#>    lax0  lax laxtot laxmax laxmin laxmax0 lax0ber laxber lax0pval laxpval
#> 1  46.8 13.0   59.8    135     55      85       1      1     0.45    0.55
#> 2   0.0  0.0    0.0     NA     NA      NA      NA     NA       NA    0.00
#> 3   5.9  2.7    8.6    140     80      90       3      3     0.92    1.00
#> 4 108.6 14.9  123.5    555     50      80       1      1     0.83    0.91
#> 5  27.2  2.1   29.3    108     45      68       3      3     0.77    0.89
#> 6 125.0 34.5  159.5    120     45      64       3      3     0.59    0.73
#>   laxfisk0 laxfisk laxfixto laxfimax laxfimin laxfmax0 laxör0 laxör laxörTOT
#> 1        0       0        0       NA       NA       NA      0   0.0      0.0
#> 2        0       0        0       NA       NA       NA      0   0.0      0.0
#> 3        0       0        0       NA       NA       NA      0   0.0      0.0
#> 4        0       0        0       NA       NA       NA      0   0.0      0.0
#> 5        0       0        0       NA       NA       NA      0   0.0      0.0
#> 6        0       0        0       NA       NA       NA      0   0.8      0.8
#>   laxörmax laxörmin mal karp mört nissö nors regnb0 regnb regnbtot regnmax0
#> 1       NA       NA   0    0    0     0    0      0     0        0       NA
#> 2       NA       NA   0    0    0     0    0      0     0        0       NA
#> 3       NA       NA   0    0    0     0    0      0     0        0       NA
#> 4       NA       NA   0    0    0     0    0      0     0        0       NA
#> 5       NA       NA   0    0    0     0    0      0     0        0       NA
#> 6      121      121   0    0    0     0    0      0     0        0       NA
#>   rödin0 rödin rödintot rödinmax rödinmin rödimax0 ruda sankr sarv sik siklö
#> 1      0     0        0       NA       NA       NA    0     0    0   0     0
#> 2      0     0        0       NA       NA       NA    0     0    0   0     0
#> 3      0     0        0       NA       NA       NA    0     0    0   0     0
#> 4      0     0        0       NA       NA       NA    0     0    0   0     0
#> 5      0     0        0       NA       NA       NA    0     0    0   0     0
#> 6      0     0        0       NA       NA       NA    0     0    0   0     0
#>   skrub rödsp småsp spigg stosp stäm sutar vimma  ål ålmax ålmin öring0 öring
#> 1     0     0     0     0     0    0     0     0 0.0    NA    NA    3.0     0
#> 2     0     0     0     0     0    0     0     0 0.0    NA    NA    0.0     0
#> 3     0     0     0     0     0    0     0     0 1.7   350   350    8.2     0
#> 4     0     0     0     0     0    0     0     0 0.0    NA    NA    2.0     0
#> 5     0     0     0     0     0    0     0     0 0.0    NA    NA    0.2     0
#> 6     0     0     0     0     0    0     0     0 0.0    NA    NA    0.7     0
#>   öringtot öringmax öringmin örinmax0 örin0ber öringber öri0pval örinpval sanst
#> 1      3.0       90       65       90        1       -9     0.48        0     0
#> 2      0.0       NA       NA       NA       NA       NA     0.00        0     0
#> 3      8.2       85       70       85        3       -9     1.00        0     0
#> 4      2.0       90       70       90        3       -9     0.99        0     0
#> 5      0.2       70       70       70        1       -9     0.86        0     0
#> 6      0.7       60       60       60        3       -9     1.00        0     0
#>   lerst smsmörb utförare ansvarig fabrikat aggregat volt strstyrk pulsfrek
#> 1     0       0   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9
#> 2     0       0   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9
#> 3     0       0   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9
#> 4     0       0   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9
#> 5     0       0   Eklövs       Sf    Lugab   Bensin  400     -9.0       -9
#> 6     0       0   Eklövs       Sf    Lugab   Bensin  500     -9.0       -9
#>   antutfis metod bredd langd lokalbre area avfiskbr medbredd medyta grumligh
#> 1        1  Kval  35.0  22.0      9.5  209        N       -9     -9        1
#> 2        1  Kval  25.0  27.0      2.0   54        N       -9     -9        1
#> 3        3 Kvant  25.0  14.7      5.0   74        N       -9     -9        1
#> 4        3 Kvant  25.0  20.0     10.0  200        N       -9     -9        1
#> 5        3 Kvant  19.8  25.0     19.8  495                -9     -9        1
#> 6        3 Kvant   5.5  25.0      5.5  138                -9     -9        1
#>   vtnfarg maxdjup medeldju bottento substr1 substr2 substr3 vattenni vattenha
#> 1       2     0.9     0.60       -9  Block1   Sten2  Block2      Låg     Strå
#> 2       2     1.1     0.90        1   Sten2  Block1   Sten1      Låg     Lugn
#> 3       2     0.4     0.35        2  Block1   Sten2   Sten1      Låg     Strö
#> 4       2     0.6     0.40        2   Sten2   Sten1  Block1      Låg     Strö
#> 5       1     0.5     0.22        2   Sten1    Grus   Sten2      Låg     Strö
#> 6       1     0.4     0.17        2   Sten2   Sten1  Block3      Låg     Strö
#>   vattente lufttemp ovegmang uvegmang uvegtyp narmiljo domtra1 domtra2 beskuggn
#> 1     13.6     13.0              Mått    Moss    Blask      Al                0
#> 2     16.6     17.0     Ring     Mått    Blom    Lövsk      Al   Björk       20
#> 3     18.3     15.0              Mått    Moss    Lövsk   Björk      Ek        0
#> 4     16.9     15.2              Mått    Palg    Lövsk     Bok      Al        0
#> 5     18.7     17.2              Mått    Moss    Lövsk      Al     Bok       50
#> 6     18.5     17.5              Mått    Moss    Lövsk      Al     Ask       90
#>   vedivant vedivpar vandhind typavpop lokalvar kalkpave kalkdatu kalkntyp
#> 1        3      1.4      Upp    Vandr        2       Ja       -9         
#> 2        0      0.0      Upp    Vandr        0       Ja       -9      Bla
#> 3        0      0.0      Upp    Vandr        1       Ja       -9         
#> 4        0      0.0      Upp    Vandr        1       Ja       -9         
#> 5        1      0.2      Upp                 2      Nej        0         
#> 6        5      3.6      Upp                 2      Nej        0         
#>   paverkt1 paverkt2 paverkt3 pastyrk1 pastyrk2 pastyrk3 elfiske_anmarkni
#> 1                                  -9       -9       -9                 
#> 2                                  -9       -9       -9                 
#> 3     Regl    Flott                 2        1       -9                 
#> 4    Bioto    Flott                 1        1       -9                 
#> 5    Rensn                          2       -9       -9                 
#> 6    Rensn                          1       -9       -9                 
#>   xkoorlok ykoorlok wgS84_Dec_N wgS84_Dec_E provplatsID stationsID
#> 1   624210   143405    56.30229    14.73989      107181     204506
#> 2   624234   143389    56.30443    14.73724      167502     234215
#> 3   624354   143339    56.31513    14.72886      167514     247355
#> 4   623295   143505    56.22027    14.75829      107049     247323
#> 5   623304   143522    56.22110    14.76100      363945     363944
#> 6   623312   143533    56.22183    14.76275      363947     363946
#>             eU_CD      mS_CD svaR_Ver kommunnr    kommun   år månad vtyP_ED
#> 1 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     9     Hav
#> 2 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     8     Hav
#> 3                 WA96144336              1082 Karlshamn 2021     9     Hav
#> 4 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     9     Hav
#> 5 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     8     Hav
#> 6 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     8     Hav
#>   vtyp_123 hoh avriomrk andsjopr avstupp avstner mindistsj lutninG_PROM
#> 1        3  56   <10000      >10    10.0     2.5       2.5          3.6
#> 2        3  57   <10000      >10    10.0     2.6       2.6          3.6
#> 3        3  61   <10000      >10    10.0     4.4       4.4          6.3
#> 4        3  28   <10000      >10     7.6     7.3       7.3          4.5
#> 5        3  25   <10000      >10     7.6     7.3       7.3          1.4
#> 6        3  26   <10000      >10     7.4     7.5       7.4          6.3
#>   medtempar medT_JULI nsp nölax pred_nölax p_VIX_nölax nandtol pred_nandtol
#> 1         7      15.5   4  62.8    40.4553      0.6767  0.0000       0.0188
#> 2         7      15.5  NA    NA    33.9831          NA      NA       0.0211
#> 3         7      15.5   6  16.8    30.0529      0.2770  0.0055       0.0196
#> 4         7      15.5   4 125.5    32.4066      0.9216  0.0044       0.0230
#> 5         7      15.5   4  29.5    35.8316      0.4205  0.0000       0.0249
#> 6         7      15.5   3 160.2    14.0679      0.9941  0.0000       0.0361
#>   p_VIX_nandtol nandlith pred_nandlith p_VIX_nandlith spproptol pred_spproptol
#> 1        0.8178   1.0000        0.9776         0.7073    0.0000         0.0684
#> 2            NA       NA        0.9783             NA        NA         0.0317
#> 3        0.6677   0.9945        0.9849         0.5703    0.1667         0.0400
#> 4        0.7136   0.9956        0.9852         0.5794    0.2500         0.0859
#> 5        0.8517   1.0000        0.9852         0.6708    0.0000         0.1205
#> 6        0.8961   1.0000        0.9851         0.6712    0.0000         0.0740
#>   p_VIX_spproptol sppropint pred_sppropintol p_VIX_sppropintol kvot pred_kvot
#> 1          0.8818    0.7500           0.5065              0.74    1    0.2373
#> 2              NA        NA           0.6540                NA   NA    0.2373
#> 3          0.1633    0.5000           0.6772              0.32    1    0.2373
#> 4          0.1557    0.5000           0.6132              0.38    1    0.2373
#> 5          0.9437    0.7500           0.4591              0.77    1    0.2373
#> 6          0.8911    0.6667           0.6728              0.49    1    0.2373
#>   p_VIX_kvot simpson pred_simpson p_VIXh_simpson  viXsm   viXh    vix viX_klass
#> 1     0.9303  0.4569       0.3637         0.7402 0.7636 0.6178 0.7923         1
#> 2         NA      NA       0.2230             NA 0.0000 0.0000 0.0000         5
#> 3     0.9303  0.5144       0.2555         0.3456 0.5253 0.4042 0.4887         2
#> 4     0.9303  0.1608       0.4005         0.3408 0.7045 0.5718 0.6145         2
#> 5     0.9303  0.5117       0.5350         0.9351 0.6997 0.5800 0.7657         1
#> 6     0.9303  0.3511       0.3455         0.9838 0.7723 0.7730 0.8127         1
#>   fisktom                                  viX_anmärkni viX_anmärkni2
#> 1      NA                                                            
#> 2       1 Klass 5 pga att fisk saknas. Kräfta fångades.              
#> 3      NA                                                            
#> 4      NA                                                            
#> 5      NA                                                            
#> 6      NA                                                            
#>   pre_sd_VIX_c    p_1    p_2    p_3    p_4    p_5 p23_diff viX23_gränsfall
#> 1       0.1436 0.6180 0.3703 0.0116 0.0002 0.0000   0.3587              NA
#> 2       0.1664 0.0000 0.0000 0.0000 0.0000 1.0000   0.0000              NA
#> 3       0.1545 0.0459 0.5107 0.3611 0.0781 0.0042   0.1496              NA
#> 4       0.1350 0.1593 0.7039 0.1310 0.0058 0.0000   0.5729              NA
#> 5       0.1198 0.5548 0.4389 0.0063 0.0000 0.0000   0.4326              NA
#> 6       0.1416 0.6731 0.3196 0.0072 0.0001 0.0000   0.3124              NA
#>   gränsfall_string viXmorf påverkat_VIXmorf
#> 1                   0.6287                0
#> 2                   0.0199                1
#> 3                   0.5305                0
#> 4                   0.4575                0
#> 5                   0.4288                0
#> 6                   0.3659                0
#>                                                  komm_VIXmorf
#> 1 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 2          VIX<0,467 och morfologiskt påverkat enligt VIXmorf
#> 3 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 4 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 5 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 6 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
dim(data)
#> [1]  60 244