Get electrofishing data from the SERS API

This function returns electrofishing data from the SERS API. You can filter the data by county (lan), municipality (kommun), main drainage area (haroNr), start date (startdatum) and end date (slutdatum). You must provide at least one of the parameters. The function does not check if the parameters are valid and it is possible to give combinations that does not make sense, e.g. a haroNr and lan that does not intersect. It is also possible send requests that return a lot of data, so be careful.

Usage

sers_vix_rapport(
  lan = NULL,
  kommun = NULL,
  haroNr = NULL,
  startdatum = NULL,
  slutdatum = NULL
)

Arguments

lan

character string with the name of a county (län) in Sweden

kommun

character string with the name of a municipality (kommun) in Sweden

haroNr

integer with the main drainage area (huvudavrinningsområde) number

startdatum

character string with the start date in the format "YYYY-MM-DD"

slutdatum

character string with the end date in the format "YYYY-MM-DD"

Value

A data frame with the electrofishing data. There are 244 columns in the data frame consult SERS-documentation.

See also

Examples

# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 86000 (Mörrumsån) in 2021
data <- sers_vix_rapport(haroNr = 86000, startdatum = "2021-01-01", slutdatum = "2021-12-31")
head(data)
#>        län länskod harO_NR huvudavrinningsområde hflodomr biflnr    vattendrag
#> 1 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 2 Blekinge      10   86000             Mörrumsån      086      S Bjällerbäcken
#> 3 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 4 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 5 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#> 6 Blekinge      10   86000             Mörrumsån      086      0     Mörrumsån
#>   xkoorvdr ykoorvdr              lokal lokalnr  rT90_X  rT90_Y swereF99_TM_N
#> 1   622563   143423     Fotbollsplan S         6252940 1429620       6250537
#> 2   622563   143423         Ladugården         6253746 1429368       6251340
#> 3   622563   143423 Pool 1 laxodlingen         6229800 1434450       6227466
#> 4   622563   143423       Pool 19 övre         6230700 1434350       6228364
#> 5   622563   143423            Flacket         6231992 1434841       6229661
#> 6   622563   143423          Rosendala         6232350 1434850       6230019
#>   swereF99_TM_E fiskedatum  syfte artantal abbor asp benlö bjöna besim stesi
#> 1        479349   20210831 Nö-lax        6     0   0   0.0     0  17.9     0
#> 2        479088   20210816    Nap        4     0   0   0.0     0   0.0     2
#> 3        484447   20210908 Nö-lax        7     0   0   0.8     0   0.0     0
#> 4        484337   20210901 Nö-lax        7     0   0   0.0     0   0.0     0
#> 5        484812   20210830   Uppf        7     0   0   1.3     0   0.0     0
#> 6        484817   20210901 Nö-lax        3     0   0   0.0     0   0.0     0
#>   simpa hosim braxe bäcne flone havne nejonöga bäcrö0 bäcrö bäcrötot bäcrömax
#> 1     0     0     0     0     0     0      0.0      0     0        0       NA
#> 2     0     0     0     0     0     0      0.0      0     0        0       NA
#> 3     0     0     0     0     0     0      0.0      0     0        0       NA
#> 4     0     0     0     0     0     0      1.9      0     0        0       NA
#> 5     0     0     0     0     0     0      2.8      0     0        0       NA
#> 6     0     0     0     0     0     0      0.0      0     0        0       NA
#>   bäcrömin bäCRmax0 bäcröxö0 bäcröxör bäcröxöt bäCXömax bäCXömin bäCXömx0 elrit
#> 1       NA       NA        0        0        0       NA       NA       NA   1.5
#> 2       NA       NA        0        0        0       NA       NA       NA   5.4
#> 3       NA       NA        0        0        0       NA       NA       NA  93.6
#> 4       NA       NA        0        0        0       NA       NA       NA 172.2
#> 5       NA       NA        0        0        0       NA       NA       NA 146.9
#> 6       NA       NA        0        0        0       NA       NA       NA  32.2
#>   flodkräf signkräf kräfta färna gers grolö grönl gädda gös harr0 harr harrtot
#> 1        0      0.0      0   0.0    0     0     0   0.0   0     0    0       0
#> 2        0      2.3      0   0.0    0     0     0   0.0   0     0    0       0
#> 3        0      0.0      0   0.0    0     0     0   0.8   0     0    0       0
#> 4        0      0.0      0  13.8    0     0     0   0.0   0     0    0       0
#> 5        0      0.0      0   2.9    0     0     0   0.0   0     0    0       0
#> 6        0      0.0      0   0.0    0     0     0   0.0   0     0    0       0
#>   harrmax harrmin harrmax0 id kanrö0 kanrö kanrötot lake  lax0  lax laxtot
#> 1      NA      NA       NA  0      0     0        0    0   5.4  0.0    5.4
#> 2      NA      NA       NA  0      0     0        0    0   0.0  0.0    0.0
#> 3      NA      NA       NA  0      0     0        0    0  50.8 10.2   61.0
#> 4      NA      NA       NA  0      0     0        0    0  96.0  7.7  103.7
#> 5      NA      NA       NA  0      0     0        0    0  34.5  0.0   34.5
#> 6      NA      NA       NA  0      0     0        0    0 137.2 14.1  151.3
#>   laxmax laxmin laxmax0 lax0ber laxber lax0pval laxpval laxfisk0 laxfisk
#> 1     85     70      85       3     -9     0.95    0.00        0       0
#> 2     NA     NA      NA      NA     NA       NA    0.00        0       0
#> 3    115     55      90       3      1     0.83    0.91        0       0
#> 4    130     40      90       3      3     0.87    0.96        0       0
#> 5     70     40      70       3     -9     0.82    0.00        0       0
#> 6    125     45      80       3      3     0.83    0.97        0       0
#>   laxfixto laxfimax laxfimin laxfmax0 laxör0 laxör laxörTOT laxörmax laxörmin
#> 1        0       NA       NA       NA    0.7     0      0.7       70       70
#> 2        0       NA       NA       NA    0.0     0      0.0       NA       NA
#> 3        0       NA       NA       NA    0.0     0      0.0       NA       NA
#> 4        0       NA       NA       NA    1.5     0      1.5       60       60
#> 5        0       NA       NA       NA    0.0     0      0.0       NA       NA
#> 6        0       NA       NA       NA    1.0     0      1.0       70       70
#>   mal karp mört nissö nors regnb0 regnb regnbtot regnmax0 rödin0 rödin rödintot
#> 1   0    0  0.7     0    0      0     0        0       NA      0     0        0
#> 2   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 3   0    0  0.7     0    0      0     0        0       NA      0     0        0
#> 4   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 5   0    0  0.0     0    0      0     0        0       NA      0     0        0
#> 6   0    0  0.0     0    0      0     0        0       NA      0     0        0
#>   rödinmax rödinmin rödimax0 ruda sankr sarv sik siklö skrub rödsp småsp spigg
#> 1       NA       NA       NA    0     0    0   0     0     0     0     0     0
#> 2       NA       NA       NA    0     2    0   0     0     0     0     0     0
#> 3       NA       NA       NA    0     0    0   0     0     0     0     0     0
#> 4       NA       NA       NA    0     0    0   0     0     0     0     0     0
#> 5       NA       NA       NA    0     0    0   0     0     0     0     0     0
#> 6       NA       NA       NA    0     0    0   0     0     0     0     0     0
#>   stosp stäm sutar vimma  ål ålmax ålmin öring0 öring öringtot öringmax
#> 1     0    0     0     0 0.0    NA    NA   25.6   9.1     34.7      115
#> 2     0    0     0     0 0.0    NA    NA    0.0   0.0      0.0       NA
#> 3     0    0     0     0 0.7   200   200    7.9   0.0      7.9       90
#> 4     0    0     0     0 0.0    NA    NA    6.0   0.0      6.0       80
#> 5     0    0     0     0 0.0    NA    NA    1.3   0.0      1.3       75
#> 6     0    0     0     0 0.0    NA    NA    0.0   0.0      0.0       NA
#>   öringmin örinmax0 örin0ber öringber öri0pval örinpval sanst lerst smsmörb
#> 1       60       95        3        3     0.92     0.89     0     0       0
#> 2       NA       NA       NA       NA     0.00     0.00     0     0       0
#> 3       60       90        3       -9     1.00     0.00     0     0       0
#> 4       60       80        1       -9     0.86     0.00     0     0       0
#> 5       75       75        3       -9     1.00     0.00     0     0       0
#> 6       NA       NA       NA       NA     0.00     0.00     0     0       0
#>   utförare ansvarig fabrikat aggregat volt strstyrk pulsfrek antutfis metod
#> 1   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 2    Lst07    Lst07  Smith-r  Batteri  305      0.6        0        3 Kvant
#> 3   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 4   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 5   Mörrum   Mörrum    Lugab   Bensin  500      0.6       -9        3 Kvant
#> 6   Mörrum      Dcf    Lugab   Bensin  500      0.6       -9        3 Kvant
#>   bredd langd lokalbre area avfiskbr medbredd medyta grumligh vtnfarg maxdjup
#> 1  20.0  17.0      8.0  136        N     -9.0     -9        1       2    0.40
#> 2   2.3  50.0      2.3  108               2.5     -9        1       3    0.37
#> 3  40.0  20.0      7.0  140        N     -9.0     -9        1       2    0.50
#> 4  25.0  17.0      8.0  136        N     -9.0     -9        1       2    0.60
#> 5  40.0  11.4      6.8   78        N     -9.0     -9        1       2    0.60
#> 6  30.0  14.7      8.0  118        N     -9.0     -9        1       2    0.50
#>   medeldju bottento substr1 substr2 substr3 vattenni vattenha vattente lufttemp
#> 1     0.20        2   Sten2   Sten1  Block1      Låg     Strö     16.6     14.6
#> 2     0.22        1    Grus   Sten1    Sand      Med     Strö     14.9     17.0
#> 3     0.30        2  Block1   Sten2  Block3      Låg     Strö     17.0     12.4
#> 4     0.30        2   Sten2   Sten1  Block1      Låg     Strö     16.9     18.0
#> 5     0.20        2    Sand    Grus   Sten1      Låg     Strö     16.5     19.0
#> 6     0.20        1   Sten1   Sten2  Block2      Låg     Strö     16.9     11.2
#>   ovegmang uvegmang uvegtyp narmiljo domtra1 domtra2 beskuggn vedivant vedivpar
#> 1              Mått    Palg    Lövsk      Al               80        0      0.0
#> 2     Sakn     Mått    Moss     Åker    Lönn      Al       50        0      0.0
#> 3              Mått    Palg    Artif                       -9       -9     -9.0
#> 4              Mått    Moss    Lövsk      Al    Lönn       20        4      2.9
#> 5     Ring     Mått    Moss      Äng      Al               10        2      2.6
#> 6              Mått    Blom    Lövsk      Al                0       -9     -9.0
#>   vandhind typavpop lokalvar kalkpave kalkdatu kalkntyp paverkt1 paverkt2
#> 1      Upp    Vandr        1       Ja       -9              Regl    Torrf
#> 2      Upp    Vandr        1      Nej        0             Jordb    Rensn
#> 3                         -9                 0                           
#> 4      Upp    Vandr        1       Ja       -9             Fledr         
#> 5     Inga    Vandr        1       Ja       -9             Fledr         
#> 6                         -9                 0                           
#>   paverkt3 pastyrk1 pastyrk2 pastyrk3                  elfiske_anmarkni
#> 1                 3        1       -9                                  
#> 2    Artif        2        2        1 5 % av ytan avräknad av utföraren
#> 3                -9       -9       -9                                  
#> 4                 2       -9       -9                                  
#> 5                 1       -9       -9                                  
#> 6                -9       -9       -9                                  
#>   xkoorlok ykoorlok wgS84_Dec_N wgS84_Dec_E provplatsID stationsID
#> 1   625294   142962    56.39899    14.66544      167582     234315
#> 2   625374   142936    56.40619    14.66114      364001     364000
#> 3   622980   143445    56.19190    14.74940      106988     247310
#> 4   623070   143435    56.19997    14.74756      107004     247313
#> 5   623199   143484    56.21164    14.75515      363941     363940
#> 6   623235   143485    56.21485    14.75521      107033     204955
#>             eU_CD      mS_CD svaR_Ver kommunnr    kommun   år månad vtyP_ED
#> 1                 WA96144336              1082 Karlshamn 2021     8     Hav
#> 2 SE625140-479092 WA64692890 SVAR2016      763  Tingsryd 2021     8        
#> 3 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     9     Hav
#> 4                 WA96144336              1082 Karlshamn 2021     9     Hav
#> 5 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     8     Hav
#> 6 SE623952-143521 WA96144336 SVAR2016     1082 Karlshamn 2021     9     Hav
#>   vtyp_123 hoh avriomrk andsjopr avstupp avstner mindistsj lutninG_PROM
#> 1        3 110   <10000      >10     0.4    10.0       0.4          5.6
#> 2        1 112     <100      <05     5.5    10.0       5.5          0.3
#> 3        3   8   <10000      >10    10.0     4.4       4.4          6.3
#> 4        3  13   <10000      >10    10.0     5.3       5.3          3.6
#> 5        3  18   <10000      >10     8.6     6.3       6.3          3.0
#> 6        3  24   <10000      >10     8.2     6.7       6.7          5.8
#>   medtempar medT_JULI nsp nölax pred_nölax p_VIX_nölax nandtol pred_nandtol
#> 1         7      15.5   5  40.1    27.5113      0.6513  0.0116       0.0508
#> 2         7      15.5   3   0.0     7.5964      0.0165  0.0000       0.0064
#> 3         7      16.5   7  68.9    38.3809      0.7291  0.0133       0.0162
#> 4         7      16.5   5 109.7    33.9831      0.8897  0.0000       0.0202
#> 5         7      16.5   6  35.8    44.9944      0.4063  0.0069       0.0179
#> 6         7      15.5   2 151.3    33.6657      0.9423  0.0000       0.0206
#>   p_VIX_nandtol nandlith pred_nandlith p_VIX_nandlith spproptol pred_spproptol
#> 1        0.7842   0.9884        0.9169         0.7483    0.2000         0.1108
#> 2        0.7013   0.7872        0.9890         0.0724    0.0000         0.0274
#> 3        0.5319   0.9819        0.9849         0.4829    0.4286         0.0571
#> 4        0.8260   1.0000        0.9857         0.6682    0.0000         0.0602
#> 5        0.6323   0.9931        0.9858         0.5530    0.1667         0.0497
#> 6        0.8285   1.0000        0.9856         0.6687    0.0000         0.0641
#>   p_VIX_spproptol sppropint pred_sppropintol p_VIX_sppropintol kvot pred_kvot
#> 1          0.2890    0.6000           0.4528              0.64    1    0.2373
#> 2          0.7714    0.3333           0.7720              0.12   NA    0.6947
#> 3          0.0172    0.2857           0.6068              0.20    1    0.2373
#> 4          0.8663    0.6000           0.6189              0.48    1    0.2373
#> 5          0.1906    0.5000           0.6895              0.31    1    0.2373
#> 6          0.8741    0.5000           0.6790              0.32    1    0.2373
#>   p_VIX_kvot simpson pred_simpson p_VIXh_simpson  viXsm   viXh    vix viX_klass
#> 1     0.9303  0.5705       0.2525         0.2492 0.7438 0.5657 0.6747         2
#> 2         NA  0.5794       0.2803         0.2829 0.0711 0.3645 0.3372         3
#> 3     0.9303  0.5419       0.3549         0.5086 0.5864 0.4510 0.4825         2
#> 4     0.9303  0.5412       0.3818         0.5747 0.7422 0.6394 0.7768         1
#> 5     0.9303  0.3667       0.3335         0.9032 0.5504 0.5807 0.5041         2
#> 6     0.9303  0.2894       0.3270         0.8866 0.7159 0.7273 0.7610         1
#>   fisktom viX_anmärkni viX_anmärkni2 pre_sd_VIX_c    p_1    p_2    p_3    p_4
#> 1      NA                                  0.1607 0.3215 0.5807 0.0914 0.0062
#> 2      NA                                  0.0983 0.0000 0.0937 0.6459 0.2557
#> 3      NA                                  0.1524 0.0400 0.5010 0.3733 0.0815
#> 4      NA                                  0.1490 0.5736 0.4077 0.0183 0.0004
#> 5      NA                                  0.1551 0.0570 0.5381 0.3359 0.0658
#> 6      NA                                  0.1460 0.5322 0.4459 0.0215 0.0004
#>      p_5 p23_diff viX23_gränsfall gränsfall_string viXmorf påverkat_VIXmorf
#> 1 0.0001   0.4893              NA                   0.7378                0
#> 2 0.0046   0.5522              NA                   0.1959                1
#> 3 0.0042   0.1278              NA                   0.7274                0
#> 4 0.0000   0.3894              NA                   0.5565                0
#> 5 0.0032   0.2022              NA                   0.4501                0
#> 6 0.0000   0.4244              NA                   0.2918                0
#>                                                  komm_VIXmorf
#> 1 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 2          VIX<0,467 och morfologiskt påverkat enligt VIXmorf
#> 3 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 4 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 5 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
#> 6 God/Hög ekologisk status enligt VIX. VIXmorf ej tillämpbart
dim(data)
#> [1]  60 244